Flytter mapping til kakfa-package
Ønsker ikke at core skal kjenne til det ytre grensesnittet - kun andre veien
